requirements.txt
pydle>=1.1.0
peewee
pyyaml
httpx
lxml
parsedatetime
pycountry
whoosh
pillow
humanize
BeautifulSoup4
pure-sasl
docutils
tornado==6.1
twitter
pathlib
https://github.com/rfk/pyenchant/tarball/master#egg=pyenchant
textblob
ccy
pymediainfo
